+++ src/Resources/contao/languages/fr/tl_layout.php
@@ -1,0 +1,27 @@
+<?php
+/**
+ * Translations are managed using Transifex. To create a new translation
+ * or to help to maintain an existing one, please register at transifex.com.
+ *
+ * @link https://www.transifex.com/signup/
+ * @link https://www.transifex.com/projects/p/contao-bootstrap/language/fr/
+ *
+ * @license http://www.gnu.org/licenses/lgpl-3.0.html LGPL
+ *
+ * last-updated: 2018-01-05T14:08:03+00:00
+ */
+
+
+$GLOBALS['TL_LANG']['tl_layout']['bootstrap_containerClass']['0']   = 'Classe du Container';
+$GLOBALS['TL_LANG']['tl_layout']['bootstrap_containerElement']['0'] = 'Élément Container';
+$GLOBALS['TL_LANG']['tl_layout']['bootstrap_footerClass']['0']      = 'Classe CSS du pied de page';
+$GLOBALS['TL_LANG']['tl_layout']['bootstrap_footerClass']['1']      = 'Peut être utilisé pour assigner la classe .container au pied de page par exemple';
+$GLOBALS['TL_LANG']['tl_layout']['bootstrap_headerClass']['0']      = 'Classe CSS de l\'entête de page';
+$GLOBALS['TL_LANG']['tl_layout']['bootstrap_headerClass']['1']      = 'Peut être utilisé pour assigner la classe .container à l\'entête de page par exemple.';
+$GLOBALS['TL_LANG']['tl_layout']['bootstrap_importFile']['0']       = 'Fichier';
+$GLOBALS['TL_LANG']['tl_layout']['bootstrap_leftClass']['0']        = 'Classe CSS de la grille de la colonne de gauche';
+$GLOBALS['TL_LANG']['tl_layout']['bootstrap_mainClass']['0']        = 'Classe CSS de la grille de la colonne principale';
+$GLOBALS['TL_LANG']['tl_layout']['bootstrap_rightClass']['0']       = 'Classe CSS de la grille de la colonne de droite';
+$GLOBALS['TL_LANG']['tl_layout']['container_legend']                = 'Container';
+$GLOBALS['TL_LANG']['tl_layout']['unsafeTpl']                       = 'Dans les autres thèmes';
+