+++ CHANGELOG.md
@@ -2,6 +2,40 @@
 Changelog
 =========
 
+2.0.4 (2019-11-15)
+------------------
+
+[Full Changelog](https://github.com/contao-bootstrap/layout/compare/2.0.3...2.0.4)
+
+### Fixed
+
+ - Add img css class to picture image instead of to attributes ([#15](https://github.com/contao-bootstrap/layout/issues/15))
+ - Do not add html5 shiv for Contao >= 4.8 ([#13](https://github.com/contao-bootstrap/layout/issues/13))
+
+2.0.3 (2018-09-06)
+------------------
+
+[Full Changelog](https://github.com/contao-bootstrap/layout/compare/2.0.2...2.0.3)
+
+ - Update translations
+ - Fix incompatibility with Contao 4.4
+
+2.0.2 (2018-08-27)
+------------------
+
+[Full Changelog](https://github.com/contao-bootstrap/layout/compare/2.0.1...2.0.2)
+
+ - Do not require that bootstrap environment is public.
+
+
+2.0.1 (2018-07-24)
+------------------
+
+[Full Changelog](https://github.com/contao-bootstrap/layout/compare/2.0.0...2.0.1)
+
+ - Update translations
+ 
+
 2.0.0 (2018-01-05)
 ------------------